Move no_std clock into std, allow it to be used everywhere
Also, rename the with_std clock to live alongside it (as
FakeRelativeClock and FakeAbsoluteClock, respectively)
